% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/su-functions.R
\name{addLabelsSUDF}
\alias{addLabelsSUDF}
\title{addLabelsSUDF}
\usage{
addLabelsSUDF(df, output, pkParameter)
}
\arguments{
\item{df}{Table of sensitivity and uncertainty analysis results.}

\item{output}{String. Output name.}

\item{pkParameter}{String. PK parameter name.}
}
\value{
Formatted table of sensitivity and uncertainty analysis results.
}
\description{
Rename headings in summary table of sensitivity and uncertainty analysis results.
}
