% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{addNewPkParameter}
\alias{addNewPkParameter}
\title{addNewPkParameter}
\usage{
addNewPkParameter(pkParameterName, standardPKParameter, startTime, endTime)
}
\arguments{
\item{pkParameterName}{The name of the PK parameter to be added.}

\item{standardPKParameter}{The standard PK parameter on which the PK parameter to be added is based, selected from among the list of strings in \code{ospsuite::allPKParameterNames()}.}

\item{startTime}{Start time within the simulation run for calculation of the PK parameter.}

\item{endTime}{End time within the simulation run for calculation of the PK parameter.}
}
\description{
Add a new PK parameter to the managed list of OSP PK parameters.
}
