% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morris-functions.R
\name{Bmat}
\alias{Bmat}
\title{Bmat}
\usage{
Bmat(k)
}
\arguments{
\item{k}{The number of parameter paths.}
}
\value{
Bmat matrix as per Morris, 1991.
}
\description{
Function to construct the Bmat matrix as per Morris, 'Factorial Sampling Plans for Preliminary Computational Experiments', 1992.
}
\keyword{internal}
