% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{buildSettingsCMD}
\alias{buildSettingsCMD}
\title{buildSettingsCMD}
\usage{
buildSettingsCMD(parameters, outputs)
}
\arguments{
\item{parameters}{A list of \code{SAParameter} objects.}

\item{outputs}{A list of \code{SAOutput} objects.}
}
\value{
A summary of all parameter and outputs selected for a sensitivity analysis
}
\description{
Build a summary of all parameter and outputs selected for a sensitivity analysis.
}
