% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{checkOutputsExistInSimulation}
\alias{checkOutputsExistInSimulation}
\title{checkOutputsExistInSimulation}
\usage{
checkOutputsExistInSimulation(
  simulation,
  outputPaths,
  simulationName,
  stopIfNotFound = FALSE
)
}
\arguments{
\item{simulation}{A PKML simulation object.}

\item{outputPaths}{A victor of strings of output paths.}

\item{simulationName}{A descriptive name for the \code{simulation}.}

\item{stopIfNotFound}{Logical value. When \code{TRUE}, results in an error when an output in \code{outputPaths} is not found in \code{simulation}.}
}
\value{
A list of logical values indicating whether each outut path in \code{outputPaths} is found in \code{simulation}.
}
\description{
Function to verify that a list of output paths exists in a simulation.
}
