% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{checkParametersExistInSimulation}
\alias{checkParametersExistInSimulation}
\title{checkParametersExistInSimulation}
\usage{
checkParametersExistInSimulation(
  simulation,
  parameterPaths,
  simulationName = NULL,
  stopIfNotFound = FALSE
)
}
\arguments{
\item{simulation}{A PKML simulation object.}

\item{simulationName}{A descriptive name for the \code{simulation}.}

\item{stopIfNotFound}{Logical value. When \code{TRUE}, results in an error when a parameter in \code{parameterPaths} is not found in \code{simulation}.}

\item{outputPaths}{A victor of strings of parameter paths.}
}
\value{
A list of logical values indicating whether each outut path in \code{parameterPaths} is found in \code{simulation}.
}
\description{
Function to verify that a list of paramter paths exists in a simulation.
}
