% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/su-functions.R
\name{cleanUpSUDf}
\alias{cleanUpSUDf}
\title{cleanUpSUDf}
\usage{
cleanUpSUDf(df)
}
\arguments{
\item{df}{Un-formatted table of sensitivity and uncertainty analysis results.}
}
\value{
Formatted table of sensitivity and uncertainty analysis results.
}
\description{
Format a summary table of sensitivity and uncertainty analysis results.
}
\keyword{internal}
