% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling-functions.R
\docType{data}
\name{distribution}
\alias{distribution}
\title{distribution}
\format{
A named \code{list} of distribution constructor functions.
}
\usage{
distribution
}
\description{
A named list of distribution constructors used to create
distribution objects for global sensitivity analyses. Each element is a
constructor function returning a distribution object: \code{Uniform},
\code{LogUniform}, \code{Normal} and \code{LogNormal}.
}
\keyword{datasets}
