% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-generation.R
\docType{data}
\name{distributionStringsFnsList}
\alias{distributionStringsFnsList}
\title{distributionStringsFnsList}
\format{
An object of class \code{list} of length 4.
}
\usage{
distributionStringsFnsList
}
\description{
Generates a string listing the values of the distribution parameters for the \code{SAParameter} object \code{parameter}
}
\keyword{internal}
