% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{error}
\alias{error}
\title{error}
\usage{
error(condition, errorMessage = NULL)
}
\arguments{
\item{condition}{A logical condition}

\item{errorMessage}{Message to be printed before stopping in case \code{condition} is \code{TRUE}}
}
\description{
Stops if \code{condition} is \code{TRUE}
}
\keyword{internal}
