% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowry-functions.R
\name{generateEFASTBarGraph}
\alias{generateEFASTBarGraph}
\title{generateEFASTBarGraph}
\usage{
generateEFASTBarGraph(gsaResultsDataframe)
}
\arguments{
\item{gsaResultsDataframe}{EFAST results returned by the \code{runEFAST} function.}
}
\value{
A list of ggplot bar graphs, one corresponding to each output path/PK parameter combination.
}
\description{
Function to generate a bar graph of EFAST sensitivity analysis results.
}
