% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowry-functions.R
\name{generateGSABarGraph}
\alias{generateGSABarGraph}
\title{generateGSABarGraph}
\usage{
generateGSABarGraph(df)
}
\arguments{
\item{df}{Sobol or EFAST results returned by the \code{runSobol} or \code{runFAST} functions, respectively.}
}
\value{
A list of ggplot bar graph plots, one corresponding to each output path/PK parameter combination.
}
\description{
Function to generate a bar graph of Sobol or EFAST sensitivity analysis results.
}
