% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowry-functions.R
\name{generateLowryPlot}
\alias{generateLowryPlot}
\title{generateLowryPlot}
\usage{
generateLowryPlot(gsaResultsDataframe)
}
\arguments{
\item{gsaResultsDataframe}{Sobol or EFAST results returned by the \code{runSobol} or \code{runFAST} functions, respectively.}
}
\value{
A list of ggplot Lowry plots, one corresponding to each output path/PK parameter combination.
}
\description{
Function to generate a Lowry plot of Sobol or EFAST sensitivity analysis results.
}
