% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-generation.R
\name{generateOutputCode}
\alias{generateOutputCode}
\title{generateOutputCode}
\usage{
generateOutputCode(path, displayName, pkParameterList)
}
\arguments{
\item{path}{The parameter path in the simulation.}

\item{displayName}{A display name for the parameter in lieu of the parameter path.}

\item{pkParameterList}{A list of PK parameter names as strings, selected from among \code{ospsuite::allPKParameterNames()}}
}
\description{
Generate code for setting up a SAOutput R6 object
}
\keyword{internal}
