% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-generation.R
\name{generateParameterCode}
\alias{generateParameterCode}
\title{generateParameterCode}
\usage{
generateParameterCode(path, displayName, unit, parameterDistribution)
}
\arguments{
\item{path}{The parameter path in the simulation.}

\item{displayName}{A display name for the parameter in lieu of the parameter path.}

\item{unit}{The units of the numerical values describing the parameter distribution, such as the mean and variance of a normal distribution.}

\item{parameterDistribution}{An SADistribution R6 object describing the distribution of the parameter.}
}
\description{
Generate code for setting up a SAParameter R6 object
}
