% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efast-functions.R
\name{generateParameterFrequenciesTotal}
\alias{generateParameterFrequenciesTotal}
\title{generateParameterFrequenciesTotal}
\usage{
generateParameterFrequenciesTotal(parameters, parameterNumber)
}
\arguments{
\item{parameters}{List of \code{SAParameter} objects.}

\item{parameterNumber}{An integer pointing to a specific parameter in the list of \code{parameters}.}
}
\value{
A vector of the frequencies associated with each parameter.
}
\description{
Generate a set of frequencies with each parameter as per Saltelli, Tarantola & Chan, 1999.
}
