% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowry-functions.R
\name{generateSobolBarGraph}
\alias{generateSobolBarGraph}
\title{generateSobolBarGraph}
\usage{
generateSobolBarGraph(gsaResultsDataframe)
}
\arguments{
\item{gsaResultsDataframe}{Sobol results returned by the \code{runSobol} function.}
}
\value{
A list of ggplot bar graphs, one corresponding to each output path/PK parameter combination.
}
\description{
Function to generate a bar graph of Sobol sensitivity analysis results.
}
