% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tornado-functions.R
\name{generateTornadoPlot}
\alias{generateTornadoPlot}
\title{generateTornadoPlot}
\usage{
generateTornadoPlot(
  sensitivityDataFrame,
  generateForUncertaintyAnalysis = FALSE
)
}
\arguments{
\item{sensitivityDataFrame}{Sensitivity or uncertainty analysis results returned by \code{runSU} function.}

\item{generateForUncertaintyAnalysis}{Logical value.  If \code{TRUE}, the tornado plots will be generated for uncertainty analysis results.}
}
\value{
A list of ggplot tornado plots results, one corresponding to each output path/PK parameter combination.
}
\description{
Function to generate a tornado plot of local sensitivity analysis or uncertainty analysis results.
}
