% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowry-functions.R
\name{getEFASTBarGraph}
\alias{getEFASTBarGraph}
\title{getEFASTBarGraph}
\usage{
getEFASTBarGraph(df, outputDisplayName, pk)
}
\arguments{
\item{df}{EFAST results dataframe.}

\item{outputDisplayName}{String. The display name of the output.}

\item{pk}{String. The name of the PK parameter}
}
\value{
A list of ggplot bar graph plots, one corresponding to each output path/PK parameter combination.
}
\description{
Function to generate a bar graph of EFAST sensitivity analysis results.
}
