% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efast-functions.R
\name{getEFASTResultsDf}
\alias{getEFASTResultsDf}
\title{getEFASTResultsDf}
\usage{
getEFASTResultsDf(ffEvaluationsList, outputs)
}
\arguments{
\item{ffEvaluationsList}{List of evaluated first order and total effect indices.}

\item{outputs}{List of \code{SAOutput} objects.}
}
\description{
Format the EFAST results into a data frame.
}
