% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-sensitivity.R
\name{getLocalSensitivity}
\alias{getLocalSensitivity}
\title{getLocalSensitivity}
\usage{
getLocalSensitivity(
  simulation,
  DDIsimulation = NULL,
  parameterPaths,
  outputs,
  variationRange = 0.1,
  numberOfSteps = 2,
  runParallel = TRUE,
  updateProgress = NULL
)
}
\arguments{
\item{simulation}{PKML simulation object.}

\item{DDIsimulation}{DDI PKML simulation object.}

\item{parameterPaths}{Paths of PKML simulation parameters to be analyzed\.}

\item{outputs}{List of \code{SAOutput} objects.}

\item{variationRange}{Variation range for sensitivity analysis.}

\item{numberOfSteps}{Number of points to evaluate model within the \code{variationRange}.}

\item{runParallel}{Logical value.  Sensitivity analysis computation is run in parallel when \code{TRUE}.}

\item{updateProgress}{Logical value.  Updates shiny app GUI with sensitivity analysis progress when \code{TRUE}.}
}
\value{
description
}
\description{
Function to run local sensitivity analysis
}
\keyword{internal}
