% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowry-functions.R
\name{getLowryPlot}
\alias{getLowryPlot}
\title{getLowryPlot}
\usage{
getLowryPlot(df, outputDisplayName, pk)
}
\arguments{
\item{df}{Sobol or EFAST results  dataframe.}

\item{outputDisplayName}{String. The display name of the output.}

\item{pk}{String. The name of the PK parameter}
}
\value{
A ggplot Lowry plot for a specific output path and PK parameter combination.
}
\description{
Function to build a Lowry plot in ggplot for a spcific output and PK parameter.
}
\keyword{internal}
