% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol-functions.R
\name{getMixedMatrices}
\alias{getMixedMatrices}
\title{runSobol}
\usage{
getMixedMatrices(A, B, parameterPaths)
}
\arguments{
\item{A}{A matrix consisting of a Sobol sequence with a number of columns equal to the number of parameter paths and a number of rows equal to the sample size.}

\item{B}{A matrix consisting of a Sobol sequence with a number of columns equal to the number of parameter paths and a number of rows equal to the sample size.}

\item{parameterPaths}{A vector of strings of parameter names.}
}
\value{
A list of matrices, one for each parameter path in \code{parameterPaths}.  The i'th matrix consists of matrix \code{B} with its i'th column replaced with the corresponding column from matrix \code{A}.
}
\description{
Function to generate points in parameter space at which simulation will be run in Sobol sensitivity analysis.
}
