% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/su-functions.R
\name{getParameterListForUncertaintyAnalysis}
\alias{getParameterListForUncertaintyAnalysis}
\title{getParameterListForUncertaintyAnalysis}
\usage{
getParameterListForUncertaintyAnalysis(
  parameterPaths,
  simulation,
  variationRange = 0.1
)
}
\arguments{
\item{parameterPaths}{Paths of the parameters in the \code{simulation}.}

\item{simulation}{The simulation object which contains parameters at paths \code{parameterPaths}.}

\item{variationRange}{The variation range used to define the log-uniform distribution of the parameters in \code{parameterPaths}.}
}
\value{
A list of \code{SAParameter} objects.
}
\description{
This function assigns a log-uniform distribution with a variation range given by \code{variationRange} for parameters with a previously unspecified distribution.
}
