% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efast-functions.R
\name{getSampleRate}
\alias{getSampleRate}
\title{getSampleRate}
\usage{
getSampleRate(parameterFrequencies)
}
\arguments{
\item{parameterFrequencies}{A vector of the frequencies associated with each parameter.}
}
\value{
A maximal rate of sampling that satisfies Nyquist-Shannon with respect to the \code{parameterFrequencies}.
}
\description{
Select a sampling rate within the parameter space as per Saltelli, Tarantola & Chan, 1999.
}
