% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efast-functions.R
\name{getSamplingHypercube}
\alias{getSamplingHypercube}
\title{getSamplingHypercube}
\usage{
getSamplingHypercube(parameterFrequencies, samplingRate)
}
\arguments{
\item{parameterFrequencies}{A vector of the frequencies associated with each parameter.}

\item{samplingRate}{Maximal rate of sampling that satisfies Nyquist-Shannon with respect to the \code{parameterFrequencies}.}
}
\value{
A hypercube of points in [0, 2*pi].
}
\description{
Build a hypercube of points at which the model will be simulated.
}
