% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/su-functions.R
\name{getSensitivityClass}
\alias{getSensitivityClass}
\title{getSensitivityClass}
\usage{
getSensitivityClass(sensitivityValue)
}
\arguments{
\item{sensitivityValue}{is a non-negative numeric value.}
}
\value{
Sensitivity class.
}
\description{
Function to classify sensitivity analysis result.
}
