% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowry-functions.R
\name{getSobolBarGraph}
\alias{getSobolBarGraph}
\title{getSobolBarGraph}
\usage{
getSobolBarGraph(df, outputDisplayName, pk)
}
\arguments{
\item{df}{Sobol results dataframe.}

\item{outputDisplayName}{String. The display name of the output.}

\item{pk}{String. The name of the PK parameter}
}
\value{
A list of ggplot bar graph plots, one corresponding to each output path/PK parameter combination.
}
\description{
Function to generate a bar graph of Sobol sensitivity analysis results.
}
\keyword{internal}
