% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/su-functions.R
\name{getSUSummaryDf}
\alias{getSUSummaryDf}
\title{getSUSummaryDf}
\usage{
getSUSummaryDf(su)
}
\arguments{
\item{su}{Sensitivity and uncertainty analysis results}
}
\value{
A list of data frames, one for each output/PK parameter combination, where each table categorizes the sensitivity and uncertainty of each parameter as high/medium/low.
}
\description{
Function to summarize and categorize sensitivity and uncertainty analysis results in a table.
}
\keyword{internal}
