% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morris-functions.R
\name{getTrajectory}
\alias{getTrajectory}
\title{getTrajectory}
\usage{
getTrajectory(
  numberOfParameters,
  numberOfGridPartitions = defaultNumberOfGridPartitions
)
}
\arguments{
\item{numberOfParameters}{The number of parameter paths.}

\item{numberOfGridPartitions}{The number of grid partitions, equivalent to parameter \code{p} in Morris, 1991.}
}
\value{
Trajectory for model evaluations through parameter space as per Morris, 1991.
}
\description{
Function to construct a trajectory for model evaluations through parameter space as per Morris, 'Factorial Sampling Plans for Preliminary Computational Experiments', 1992.
}
\keyword{internal}
