% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty-analysis.R
\name{getUncertaintyAnalysisResults}
\alias{getUncertaintyAnalysisResults}
\title{getUncertaintyAnalysisResults}
\usage{
getUncertaintyAnalysisResults(
  simulation,
  DDIsimulation = NULL,
  parameters,
  outputs,
  numberOfUncertaintyAnalysisSamples,
  sensitiveParameterPaths,
  runParallel = TRUE,
  updateProgress = NULL
)
}
\arguments{
\item{simulation}{PKML simulation object.}

\item{DDIsimulation}{DDI PKML simulation object.}

\item{parameters}{List of \code{SAParameter} objects.}

\item{outputs}{List of \code{SAOutput} objects.}

\item{numberOfUncertaintyAnalysisSamples}{Number of samples at which to evaluate the simulation for the uncertainty analysis.}

\item{sensitiveParameterPaths}{Paths of \code{simulation} parameters that are deemed sensitive.}

\item{runParallel}{Logical value.  Uncertainty analysis computation is run in parallel when \code{TRUE}.}

\item{updateProgress}{Logical value.  Updates shiny app GUI with uncertainty analysis progress when \code{TRUE}.}
}
\value{
description
}
\description{
Function to run uncertainty analysis
}
