% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/su-functions.R
\name{getUncertaintyClass}
\alias{getUncertaintyClass}
\title{getUncertaintyClass}
\usage{
getUncertaintyClass(uncertaintyValue)
}
\arguments{
\item{uncertaintyValue}{is a non-negative numeric value.}
}
\value{
Uncertainty class.
}
\description{
Function to classify uncertainty analysis result.
}
