% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{\%||\%}
\alias{\%||\%}
\title{\verb{\%||\%}}
\usage{
lhs \%||\% rhs
}
\arguments{
\item{lhs}{argument}

\item{rhs}{argument}
}
\value{
\code{lhs} if not null, otherwise \code{rhs}
}
\description{
Function to select RHS argument in case LHS is NULL.  See purrr::\%||\%
}
