% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morris-functions.R
\name{J1mat}
\alias{J1mat}
\title{J1mat}
\usage{
J1mat(k)
}
\arguments{
\item{k}{The number of parameter paths.}
}
\value{
J1mat matrix as per Morris, 1991.
}
\description{
Function to construct the J1mat matrix as per Morris, 'Factorial Sampling Plans for Preliminary Computational Experiments', 1992.
}
