% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morris-functions.R
\name{Jmat}
\alias{Jmat}
\title{Jmat}
\usage{
Jmat(k)
}
\arguments{
\item{k}{The number of parameter paths.}
}
\value{
Jmat matrix as per Morris, 1991.
}
\description{
Function to construct the Jmat matrix as per Morris, 'Factorial Sampling Plans for Preliminary Computational Experiments', 1992.
}
