% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{LogNormalDistribution}
\alias{LogNormalDistribution}
\title{LogNormalDistribution}
\description{
R6 class, child class of \code{SADistribution}, defining a statistical lognormal distribution object
}
\section{Super class}{
\code{\link[ospsuite.globalsensitivity:SADistribution]{ospsuite.globalsensitivity::SADistribution}} -> \code{LogNormalDistribution}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{type}}{is the type of the probability distribution.}

\item{\code{mean}}{The mean of the lognormal distribution.}

\item{\code{CV}}{The coefficient of variation of the lognormal distribution.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LogNormalDistribution-new}{\code{LogNormalDistribution$new()}}
\item \href{#method-LogNormalDistribution-quantilesToSample}{\code{LogNormalDistribution$quantilesToSample()}}
\item \href{#method-LogNormalDistribution-clone}{\code{LogNormalDistribution$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogNormalDistribution-new"></a>}}
\if{latex}{\out{\hypertarget{method-LogNormalDistribution-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{LogNormalDistribution} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogNormalDistribution$new(mean, CV)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mean}}{The mean of the lognormal distribution.}

\item{\code{CV}}{The coefficient of variation of the lognormal distribution.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of the \code{LogNormalDistribution} class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogNormalDistribution-quantilesToSample"></a>}}
\if{latex}{\out{\hypertarget{method-LogNormalDistribution-quantilesToSample}{}}}
\subsection{Method \code{quantilesToSample()}}{
Maps a vector of \code{quantiles} to corresponding values from the lognormal distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogNormalDistribution$quantilesToSample(quantiles)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{quantiles}}{Vector of quantiles.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of values from the lognormal distribution corresponding to the input \code{quantiles}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogNormalDistribution-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LogNormalDistribution-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogNormalDistribution$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
