% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{LogUniformDistribution}
\alias{LogUniformDistribution}
\title{LogUniformDistribution}
\description{
R6 class, child class of \code{SADistribution}, defining a statistical loguniform distribution object
}
\section{Super class}{
\code{\link[ospsuite.globalsensitivity:SADistribution]{ospsuite.globalsensitivity::SADistribution}} -> \code{LogUniformDistribution}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{type}}{is the type of the probability distribution.}

\item{\code{minimum}}{The maximum of the loguniform distribution.}

\item{\code{maximum}}{The maximum of the loguniform distribution.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LogUniformDistribution-new}{\code{LogUniformDistribution$new()}}
\item \href{#method-LogUniformDistribution-quantilesToSample}{\code{LogUniformDistribution$quantilesToSample()}}
\item \href{#method-LogUniformDistribution-clone}{\code{LogUniformDistribution$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogUniformDistribution-new"></a>}}
\if{latex}{\out{\hypertarget{method-LogUniformDistribution-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{LogUniformDistribution} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogUniformDistribution$new(minimum, maximum)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{minimum}}{The minimum of the loguniform distribution.}

\item{\code{maximum}}{The maximum of the loguniform distribution.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of the \code{LogUniformDistribution} class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogUniformDistribution-quantilesToSample"></a>}}
\if{latex}{\out{\hypertarget{method-LogUniformDistribution-quantilesToSample}{}}}
\subsection{Method \code{quantilesToSample()}}{
Maps a vector of \code{quantiles} to corresponding values from the loguniform distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogUniformDistribution$quantilesToSample(quantiles)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{quantiles}}{Vector of quantiles.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of values from the loguniform distribution corresponding to the input \code{quantiles}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogUniformDistribution-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LogUniformDistribution-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogUniformDistribution$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
