% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efast-functions.R
\name{mapHypercubeToParameterSpace}
\alias{mapHypercubeToParameterSpace}
\title{mapHypercubeToParameterSpace}
\usage{
mapHypercubeToParameterSpace(parameters, hypercube)
}
\arguments{
\item{parameters}{List of \code{SAParameter} objects.}

\item{hypercube}{A hypercube of points in the percentile space (0,1) that is to be mapped onto the domain of each parameter's distribution.}
}
\value{
A list of points in the domain of each parameter at which the model is to be evaluated.
}
\description{
Use parameter distributions to map points in the sampling hypercube to values within the domain of each parameter's distribution.
}
