% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{NormalDistribution}
\alias{NormalDistribution}
\title{NormalDistribution}
\description{
R6 class, child class of \code{SADistribution}, defining a statistical normal distribution object
}
\section{Super class}{
\code{\link[ospsuite.globalsensitivity:SADistribution]{ospsuite.globalsensitivity::SADistribution}} -> \code{NormalDistribution}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{type}}{is the type of the probability distribution.}

\item{\code{mean}}{The mean of the normal distribution.}

\item{\code{stdv}}{The standard deviation of the normal distribution.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NormalDistribution-new}{\code{NormalDistribution$new()}}
\item \href{#method-NormalDistribution-quantilesToSample}{\code{NormalDistribution$quantilesToSample()}}
\item \href{#method-NormalDistribution-clone}{\code{NormalDistribution$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NormalDistribution-new"></a>}}
\if{latex}{\out{\hypertarget{method-NormalDistribution-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{NormalDistribution} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalDistribution$new(mean, stdv)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mean}}{The mean of the normal distribution.}

\item{\code{stdv}}{The standard deviation of the normal distribution.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of the \code{NormalDistribution} class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NormalDistribution-quantilesToSample"></a>}}
\if{latex}{\out{\hypertarget{method-NormalDistribution-quantilesToSample}{}}}
\subsection{Method \code{quantilesToSample()}}{
Maps a vector of \code{quantiles} to corresponding values from the normal distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalDistribution$quantilesToSample(quantiles)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{quantiles}}{Vector of quantiles.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of values from the normal distribution corresponding to the input \code{quantiles}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NormalDistribution-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NormalDistribution-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NormalDistribution$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
