% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuite.globalsensitivity-package.R
\docType{package}
\name{ospsuite.globalsensitivity-package}
\alias{ospsuite.globalsensitivity}
\alias{ospsuite.globalsensitivity-package}
\title{ospsuite.globalsensitivity: Open Systems Pharmacology Global Sensitivity Package}
\description{
A global sensitivity analysis for Open Systems Pharmacology models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Open-Systems-Pharmacology/OSPSuite.GlobalSensitivity}
  \item \url{https://www.open-systems-pharmacology.org/OSPSuite.GlobalSensitivity/}
  \item Report bugs at \url{https://github.com/Open-Systems-Pharmacology/OSPSuite.GlobalSensitivity/issues}
}

}
\author{
\strong{Maintainer}: Abdullah Hamadeh \email{ahamadeh@systemsinsilico.com}

}
\keyword{internal}
