% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efast-functions.R
\name{perturbHypercube}
\alias{perturbHypercube}
\title{perturbHypercube}
\usage{
perturbHypercube(hyperCube)
}
\arguments{
\item{hyperCube}{Hypercube of points in parameter space at which model will be evaluated.}
}
\value{
A perturbation to the points of the original \code{hypercube}.
}
\description{
Perturb the points within the sampling hypercube as per Saltelli, Tarantola & Chan, 1999.
}
\keyword{internal}
