% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morris-functions.R
\name{Pmat}
\alias{Pmat}
\title{Pmat}
\usage{
Pmat(k)
}
\arguments{
\item{k}{The number of parameter paths.}
}
\value{
Pmat matrix as per Morris, 1991.
}
\description{
Function to construct the Pmat matrix as per Morris, 'Factorial Sampling Plans for Preliminary Computational Experiments', 1992.
}
\keyword{internal}
