% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/su-functions.R
\name{renameSUDf}
\alias{renameSUDf}
\title{renameSUDf}
\usage{
renameSUDf(df)
}
\arguments{
\item{df}{Un-formatted table of sensitivity and uncertainty analysis results.}
}
\value{
Formatted table of sensitivity and uncertainty analysis results.
}
\description{
Rename headings in summary table of sensitivity and uncertainty analysis results.
}
\keyword{internal}
