% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efast-functions.R
\name{runEFAST}
\alias{runEFAST}
\title{runEFAST}
\usage{
runEFAST(
  simulation,
  DDIsimulation = NULL,
  parameters,
  outputs,
  runParallel = TRUE,
  numberOfResamples = 1,
  updateProgress = NULL,
  saveResults = FALSE,
  saveFolder = NULL,
  saveFileName = NULL
)
}
\arguments{
\item{simulation}{PKML simulation object.}

\item{DDIsimulation}{DDI PKML simulation object.}

\item{parameters}{List of \code{SAParameter} objects.}

\item{outputs}{List of \code{SAOutput} objects.}

\item{runParallel}{Logical value.  EFAST sensitivity computation is run in parallel when \code{TRUE}.}

\item{numberOfResamples}{Number of times to run the EFAST algorithm steps with resampling, as per Saltelli, Tarantola & Chan, 1999.}

\item{saveResults}{Logical value.  Results are saved if \code{TRUE}.}

\item{saveFolder}{Folder in which results will be saved if \code{saveResults} is set to \code{TRUE}.}

\item{saveFileName}{File name to which results will be saved if \code{saveResults} is set to \code{TRUE}.}
}
\description{
Run the EFAST algorithm.
}
