% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efast-functions.R
\name{runFFT2}
\alias{runFFT2}
\title{runFFT2}
\usage{
runFFT2(
  outputs,
  pkEvaluationsList,
  parameters,
  fftStructure,
  allFrequencies,
  parameterFrequencies,
  addHarmonicsForParameterNumber
)
}
\arguments{
\item{outputs}{List of \code{SAOutput} objects.}

\item{pkEvaluationsList}{List of evaluated PK parameters for each output.}

\item{parameters}{List of \code{SAParameter} objects.}

\item{fftStructure}{An empty structured list to be updated with the first order and total effects resulting from the FFT evaluations.}

\item{allFrequencies}{A vector of all integer frequencies from 0 up to the sampling rate that was used in generating the hypercube over which the pkEvaluationsList was evaluated.}

\item{parameterFrequencies}{A vector of the frequencies associated with each parameter.}

\item{addHarmonicsForParameterNumber}{Parameter number for which first order index is to be calculated}
}
\value{
First order and total effect indices of EFAST evaluation.
}
\description{
Run the Fast Fourier Transform and the sensitivity indices.
}
\keyword{internal}
