% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol-functions.R
\name{runSobol}
\alias{runSobol}
\title{runSobol}
\usage{
runSobol(
  simulation,
  DDIsimulation = NULL,
  parameters,
  outputs,
  numberOfSamples,
  runParallel = TRUE,
  updateProgress = NULL,
  saveResults = FALSE,
  saveFolder = NULL,
  saveFileName = NULL
)
}
\arguments{
\item{simulation}{PKML simulation object.}

\item{DDIsimulation}{DDI PKML simulation object.}

\item{parameters}{List of \code{SAParameter} objects.}

\item{outputs}{List of \code{SAOutput} objects.}

\item{numberOfSamples}{The number of samples in parameter space at which to evaluate the simulation.}

\item{runParallel}{Logical value.  Sobol computation is run in parallel when \code{TRUE}.}

\item{updateProgress}{Logical value.  Updates shiny app GUI with Sobol algorithm progress when \code{TRUE}.}

\item{saveResults}{Logical value.  If \code{TRUE}, the results will be saved.}

\item{saveFolder}{String indicating the path to the folder in which the results are to be saved.}

\item{saveFileName}{String indicating the file name to use when saving the results.}
}
\value{
Sobol sensitivity analysis results.
}
\description{
Function to run Sobol sensitivity analysis.
}
