% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/su-functions.R
\name{runSU}
\alias{runSU}
\title{runSU}
\usage{
runSU(
  simulation,
  DDIsimulation = NULL,
  customParameters = NULL,
  outputs,
  evaluateForAllParameters = FALSE,
  variationRange = 0.1,
  numberOfSensitivityAnalysisSteps = 2,
  sensitivityThreshold = 0.1,
  runUncertaintyAnalysis = TRUE,
  runUncertaintlyOnlyForSensitiveParameters = TRUE,
  numberOfUncertaintyAnalysisSamples = 100,
  quantiles = c(0.05, 0.5, 0.95),
  saveResults = FALSE,
  saveFolder = NULL,
  saveFileName = NULL,
  runParallel = TRUE,
  updateProgressSensitivity = NULL,
  updateProgressUncertainty = NULL
)
}
\arguments{
\item{simulation}{PKML simulation object.}

\item{DDIsimulation}{DDI PKML simulation object.}

\item{customParameters}{List of \code{SAParameter} objects with user-specified distributions.}

\item{outputs}{List of \code{SAOutput} objects.}

\item{evaluateForAllParameters}{Logical value.  If \code{TRUE}, all model parameters will be evaluated for sensitivity and those with a sensitivity greater than \code{sensitivityThreshold} will undergo uncertainty analysis.  Any parameters not included in the \code{customParameters} list will have a log-uniform distribution with bounds set by \code{variationRange}.
Sensitivity analysis settings:}

\item{variationRange}{The variation range used to define the log-uniform distribution of the parameters in \code{parameterPaths}.}

\item{numberOfSensitivityAnalysisSteps}{Positive integer.  Number of steps at which the model is simulated within a parameter's \code{variationRange} during the sensitivity analysis.}

\item{sensitivityThreshold}{Positive numeric value.  The threshold sensitivity above which a parameter will undergo uncertainty analysis if \code{runUncertaintyAnalysis} and \code{runUncertaintlyOnlyForSensitiveParameters} are both \code{TRUE}.
Uncertainty analysis settings:}

\item{runUncertaintyAnalysis}{Logical value.  If \code{TRUE}, uncertainty analysis will run.}

\item{runUncertaintlyOnlyForSensitiveParameters}{Logical value.  If \code{TRUE}, uncertainty analysis will only be run for the parameters for which sensitivity exceeds \code{sensitivityThreshold}.}

\item{numberOfUncertaintyAnalysisSamples}{Positive integer value giving the number of Monte Carlo runs for each parameter.}

\item{quantiles}{Vector of numerical values between 0 and 1.  The \code{quantiles} input sets the percentiles of the PK parameters to be evaluated in addition to the 50\% and 95\% percentiles, which are computed by default.
Save settings:}

\item{saveResults}{Logical value.  If \code{TRUE}, the results will be saved.}

\item{saveFolder}{String indicating the path to the folder in which the results are to be saved.}

\item{saveFileName}{String indicating the file name to use when saving the results.
Parallel run and updating settings}

\item{runParallel}{Logical value.  Sensitivity computation is run in parallel when \code{TRUE}.}

\item{updateProgressSensitivity}{Logical value.  Updates shiny app GUI with sensitivity analysis progress when \code{TRUE}.}

\item{updateProgressUncertainty}{Logical value.  Updates shiny app GUI with uncertainty analysis progress when \code{TRUE}.}
}
\value{
Sensitivity and uncertainty analysis results.
}
\description{
Function to run sensitivity analysis and then run uncertainy analysis for parameters deemed sensitive.
Main parameters:
}
