% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{SAOutput}
\alias{SAOutput}
\title{SAOutput}
\description{
R6 class defining a model output object
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{is the path to the output within the \code{simulation} PKML simulation.}

\item{\code{displayName}}{is the display name for the output in lieu of the output path in the \code{simulation} PKML simulation.}

\item{\code{pkParameterList}}{The list of PK parameters for this instance of the \code{SAOutput}.}

\item{\code{dimension}}{The dimension of this instance of the \code{SAOutput}.}

\item{\code{unit}}{The display unit of this instance of the \code{SAOutput}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SAOutput-new}{\code{SAOutput$new()}}
\item \href{#method-SAOutput-addPKParameter}{\code{SAOutput$addPKParameter()}}
\item \href{#method-SAOutput-clone}{\code{SAOutput$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SAOutput-new"></a>}}
\if{latex}{\out{\hypertarget{method-SAOutput-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{SAOutput} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SAOutput$new(simulation, path, displayName = NULL, unit = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{simulation}}{simulation A PKML simulation in which the parameter exists.}

\item{\code{path}}{A PKML simulation in which the output exists.}

\item{\code{displayName}}{A shorthand string for the output that substitutes for the path for display purposes.}

\item{\code{unit}}{A valid OSP unit that will be the display unit for this instance of the \code{SAOutput} object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of the \code{SAOutput} class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SAOutput-addPKParameter"></a>}}
\if{latex}{\out{\hypertarget{method-SAOutput-addPKParameter}{}}}
\subsection{Method \code{addPKParameter()}}{
Add PK parameter to be evaluated for the \code{SAOutput} instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SAOutput$addPKParameter(
  standardPKParameter,
  pkParameterDisplayName = NULL,
  startTime = NULL,
  endTime = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{standardPKParameter}}{A standard PK parameter selected from among the list of strings in \code{ospsuite::allPKParameterNames()}.}

\item{\code{pkParameterDisplayName}}{A shorthand string for the parameter that substitutes for the path for display purposes.}

\item{\code{startTime}}{Start time within the simulation run for calculation of the PK parameter.}

\item{\code{endTime}}{End time within the simulation run for calculation of the PK parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SAOutput-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SAOutput-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SAOutput$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
