% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{SAParameter}
\alias{SAParameter}
\title{SAParameter}
\description{
R6 class defining a parameter object
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{is the path to the parameter within the \code{simulation} PKML simulation.}

\item{\code{displayName}}{is the display name for the parameter in lieu of the parameter path in the \code{simulation} PKML simulation.}

\item{\code{dimension}}{is the dimension of the parameter in the \code{simulation} PKML simulation.}

\item{\code{unit}}{is the unit of the parameter in the \code{simulation} PKML simulation.}

\item{\code{distribution}}{is the distribution of the parameter}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SAParameter-new}{\code{SAParameter$new()}}
\item \href{#method-SAParameter-clone}{\code{SAParameter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SAParameter-new"></a>}}
\if{latex}{\out{\hypertarget{method-SAParameter-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{SAParameter} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SAParameter$new(
  simulation,
  path,
  displayName = NULL,
  unit = NULL,
  parameterDistribution = NULL,
  defaultVariationRangeForLogUniformDistributions = 0.1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{simulation}}{simulation A PKML simulation in which the parameter exists.}

\item{\code{path}}{A PKML simulation in which the parameter exists.}

\item{\code{displayName}}{A shorthand string for the parameter that substitutes for the path for display purposes.}

\item{\code{unit}}{A valid OSP unit used to interpret the numerical values that are input into the \code{parameterDistribution} object, such as the mean and variance of a normal distribution.}

\item{\code{parameterDistribution}}{A \code{SADistribution} object specifying the probability distribution of the parameter.}

\item{\code{defaultVariationRangeForLogUniformDistributions}}{When no \code{parameterDistribution} is specified, a loguniform distribution is assumed with multiplicative variation range given by \code{defaultVariationRangeForLogUniformDistributions}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of the \code{SAParameter} class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SAParameter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SAParameter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SAParameter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
