% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{UniformDistribution}
\alias{UniformDistribution}
\title{UniformDistribution}
\description{
R6 class, child class of \code{SADistribution}, defining a statistical uniform distribution object
}
\section{Super class}{
\code{\link[ospsuite.globalsensitivity:SADistribution]{ospsuite.globalsensitivity::SADistribution}} -> \code{UniformDistribution}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{type}}{is the type of the probability distribution.}

\item{\code{minimum}}{The minimum of the uniform distribution.}

\item{\code{maximum}}{The maximum of the uniform distribution.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UniformDistribution-new}{\code{UniformDistribution$new()}}
\item \href{#method-UniformDistribution-quantilesToSample}{\code{UniformDistribution$quantilesToSample()}}
\item \href{#method-UniformDistribution-clone}{\code{UniformDistribution$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UniformDistribution-new"></a>}}
\if{latex}{\out{\hypertarget{method-UniformDistribution-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{UniformDistribution} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UniformDistribution$new(minimum, maximum)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{minimum}}{The minimum of the uniform distribution.}

\item{\code{maximum}}{The maximum of the uniform distribution.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of the \code{UniformDistribution} class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UniformDistribution-quantilesToSample"></a>}}
\if{latex}{\out{\hypertarget{method-UniformDistribution-quantilesToSample}{}}}
\subsection{Method \code{quantilesToSample()}}{
Maps a vector of \code{quantiles} to corresponding values from the uniform distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UniformDistribution$quantilesToSample(quantiles)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{quantiles}}{Vector of quantiles.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of values from the uniform distribution corresponding to the input \code{quantiles}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UniformDistribution-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UniformDistribution-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UniformDistribution$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
