% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol-functions.R
\name{varcalc}
\alias{varcalc}
\title{varcalc}
\usage{
varcalc(fU_list, outputList)
}
\arguments{
\item{fU_list}{A list of model evaluations for different Sobol sequence matrices.  Evaluations correspond to each output/PK parameter combination.}

\item{outputList}{A vector of strings of parameter names.}
}
\value{
First order and total effect Sobol indices for each parameter path/output path/PK parameter combination.
}
\description{
Function to calculate Sobol indices from simulated model evaluations.
}
