% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-generation.R
\name{writeEFASTSensitivityFunctionToFile}
\alias{writeEFASTSensitivityFunctionToFile}
\title{writeEFASTSensitivityFunctionToFile}
\usage{
writeEFASTSensitivityFunctionToFile(numberOfReSamplesEFAST)
}
\arguments{
\item{numberOfReSamplesEFAST}{Number of times to run the EFAST algorithm.}
}
\description{
Write to file code for generating an EFAST sensitivity analyses.
}
