% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-generation.R
\name{writeMorrisSensitivityFunctionToFile}
\alias{writeMorrisSensitivityFunctionToFile}
\title{writeMorrisSensitivityFunctionToFile}
\usage{
writeMorrisSensitivityFunctionToFile(numberOfSamples)
}
\arguments{
\item{numberOfSamples}{Number of samples to run in Morris algorithm}
}
\description{
Write to file code for generating an Morris sensitivity analyses.
}
\keyword{internal}
