% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-generation.R
\name{writeParametersAndOutputsToFile}
\alias{writeParametersAndOutputsToFile}
\title{writeParametersAndOutputsToFile}
\usage{
writeParametersAndOutputsToFile(parametersList, outputsList)
}
\arguments{
\item{parametersList}{A list of SAParameter R6 objects.}

\item{outputsList}{A list of SAOutput R6 objects.}
}
\description{
Writes code for generating the lists of parameters and outputs to file.
}
\keyword{internal}
