% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-generation.R
\name{writeSobolSensitivityFunctionToFile}
\alias{writeSobolSensitivityFunctionToFile}
\title{writeSobolSensitivityFunctionToFile}
\usage{
writeSobolSensitivityFunctionToFile(numberOfSamplesGSA)
}
\arguments{
\item{numberOfSamplesGSA}{Number of sampls to test in the parameter space.}
}
\description{
Write to file code for generating a Sobol sensitivity analyses.
}
\keyword{internal}
