% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-generation.R
\name{writeSUFunctionToFile}
\alias{writeSUFunctionToFile}
\title{writeSUFunctionToFile}
\usage{
writeSUFunctionToFile(
  runUncertaintyCheckbox,
  runSensitivityAnalysisForAllParameters,
  sensitivityThreshold,
  variationRangeNumericInput,
  numberOfStepsNumericInput,
  quantilesTestInput,
  numberOfUncertaintySamples,
  runParallel
)
}
\arguments{
\item{runUncertaintyCheckbox}{Logical parameter setting whether to run uncertainty analysis.}

\item{runSensitivityAnalysisForAllParameters}{Logical parameter setting whether to run sensitivity analysis for all constant parameters in the model.}

\item{sensitivityThreshold}{Minimum sensitivity threshold needed for a parameter sensitivity to be included in results.}

\item{variationRangeNumericInput}{Range over which the sensitivity is to be assessed.}

\item{numberOfStepsNumericInput}{Number of steps within the variation range at which to assess sensitivity.}

\item{quantilesTestInput}{Quantiles at which to assess percentiles of PK parameters in the uncertainty analysis.}

\item{numberOfUncertaintySamples}{Number of points at which to simulate the model in the uncertainty analysis}

\item{runParallel}{Logical parameter setting whether to run the local sensitivity and uncertainty analyses in parallel.}
}
\description{
Write to file code for generating a local sensitivity and uncertainty analyses.
}
