% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efast-functions.R
\name{extractPKParametersFromBatchSimulationResults}
\alias{extractPKParametersFromBatchSimulationResults}
\title{extractPKParametersFromBatchSimulationResults}
\usage{
extractPKParametersFromBatchSimulationResults(
  batchSimulationResults,
  DDIbatchSimulationResults,
  outputs,
  pkEvaluationsList
)
}
\arguments{
\item{batchSimulationResults}{for the main \code{simulation}}

\item{DDIbatchSimulationResults}{for the DDI \code{simulation}}

\item{outputs}{List of \code{SAOutput} objects.}

\item{pkEvaluationsList}{An empty structured list to be updated with the PK parameter evaluations.}
}
\value{
Evaluation of PK parameters from batch simulation results.
}
\description{
Evaluate the PK parameters corresponding to each output in the simulation results obtained from batch mode simulation.
}
\keyword{internal}
