% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowry-functions.R
\name{generateGSABarGraph}
\alias{generateGSABarGraph}
\title{generateGSABarGraph}
\usage{
generateGSABarGraph(gsaResultsDataframe, plottingFunction)
}
\arguments{
\item{gsaResultsDataframe}{Sobol or EFAST results returned by the \code{runSobol} or \code{runEFAST} functions, respectively.}

\item{plottingFunction}{The name of the function to use for plotting. Choose either \code{getSobolBarGraph} for Sobol results or \code{getEFASTBarGraph} for EFAST results.}
}
\value{
A list of ggplot bar graphs, one corresponding to each output path/PK parameter combination.
}
\description{
Function to generate a bar graph of Sobol or EFAST sensitivity analysis results.
}
\keyword{internal}
