% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morris-functions.R
\name{generateMorrisPlot}
\alias{generateMorrisPlot}
\title{generateMorrisPlot}
\usage{
generateMorrisPlot(morrisResults, logPlot = FALSE)
}
\arguments{
\item{morrisResults}{Morris sensitivity results returned by \code{runMorris} function.}

\item{logPlot}{Logical setting.  The Morris results are plotted on a logarithmic scale if \code{TRUE}.}
}
\value{
A list of ggplots of Morris sensitivity analysis results, one corresponding to each output path/PK parameter combination.
}
\description{
Function to generate a plot of Morris sensitivity analysis.
}
