% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morris-functions.R
\name{runMorris}
\alias{runMorris}
\title{runMorris}
\usage{
runMorris(
  simulation,
  DDIsimulation = NULL,
  parameters,
  outputs,
  numberOfSamples,
  runParallel = TRUE,
  updateProgress = NULL,
  saveResults = FALSE,
  saveFolder = NULL,
  saveFileName = NULL
)
}
\arguments{
\item{simulation}{PKML simulation object.}

\item{DDIsimulation}{DDI PKML simulation object.}

\item{parameters}{List of \code{SAParameter} objects.}

\item{outputs}{List of \code{SAOutput} objects.}

\item{numberOfSamples}{The number of runs of the Morris algorithm to perform.}

\item{runParallel}{Logical value.  Morris computation is run in parallel when \code{TRUE}.}

\item{updateProgress}{Logical value.  Updates shiny app GUI with Morris algorithm progress when \code{TRUE}.}

\item{saveResults}{Logical value.  If \code{TRUE}, the results will be saved.}

\item{saveFolder}{String indicating the path to the folder in which the results are to be saved.}

\item{saveFileName}{String indicating the file name to use when saving the results.}
}
\value{
Morris sensitivity analysis results.
}
\description{
Function to run Morris sensitivity analysis.
}
