% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{SADistribution}
\alias{SADistribution}
\title{SADistribution}
\description{
R6 class defining a statistical distribution object
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SADistribution-clone}{\code{SADistribution$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SADistribution-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SADistribution-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SADistribution$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
